requirements.txt
apscheduler>=3.10
ollama>=0.2.0
pypdf>=5.0.0
pyyaml>=6.0
rich>=13.7
